Added error handling for Prune, Search, and Subgroup, and added check for getPrunSized
Previously, if getPrunSized was larger than the number of possible states (for example, a size of 1,000,000 pruning <R2, U> which only has 14,400 states) then it would get stuck in an infinite loop. Now, at any point, if the number of elements added to the prune table decreases, the pruning stops (because it has most likely visited over half of the possible positions).